requirements.txt
PyQt6
scapy
psutil
psycopg2-binary
networkx
matplotlib
reportlab
python-dotenv
